rollack cary

+++ src/crypto/base58.cpp
@@ -74,7 +74,7 @@ static bool DecodeBase58(const char *psz, std::vector<unsigned char> &vch,
                 "mapBase58.size() should be 256"); // guarantee not out of range
   while (*psz && !IsSpace(*psz)) {
     // Decode base58 character
-    size_t carry = mapBase58[(uint8_t)*psz];
+    int carry = mapBase58[(uint8_t)*psz];
     if (carry == -1) // Invalid b58 character
       return false;
     size_t i = 0;
